+++ packages/boethius/examples/score.js
@@ -193,16 +193,12 @@ const examples = {
         // create voices
         var soprano = scored.voice({name: "treble"}, [
             // slur across systems
-            n({value: 2, pitch: 'a4'}), n({value: 2, pitch: 'a4', slur: 1}), n({value: 1, pitch: 'ab4', slur: 1}), n({value: 1, pitch: 'c5', slur: 2}),
-            n({value: 2, pitch: 'a4', slur: 2}), n({value: 2, pitch: 'a4', slur: 1}), n({value: 1, pitch: 'ab4', slur: 1}), n({value: 1, pitch: 'c5', slur: 2}),
-            n({value: 2, pitch: 'a4', slur: 2}), n({value: 2, pitch: 'c5', slur: 1}), n({value: 1, pitch: 'ab4', slur: 1}), n({value: 1, pitch: 'c5', slur: 2}),
-            n({value: 2, pitch: 'a4', slur: 2}), n({value: 2, pitch: 'c5', slur: 1}), n({value: 1, pitch: 'ab4', slur: 1}), n({value: 1, pitch: 'c5'})
+            n({value: 2, pitch: 'a4', slur: 1}), n({value: 2, pitch: 'a4', slur: 1}),
+            r({value: 4}), n({value: 4, pitch: 'c5', slur: 2}), n({value: 4, pitch: 'c5', slur: 2}), n({value: 4, pitch: 'c5', slur: 2}),
         ]);
         var bass = scored.voice({name: "bass"}, [
-            n({value: 1, pitch: 'e3'}), n({value: 1, pitch: 'e3'}), n({value: 1, pitch: 'e3'}),
-            n({value: 1, pitch: 'e3'}), n({value: 1, pitch: 'e3'}), n({value: 1, pitch: 'e3'}),
-            n({value: 1, pitch: 'e3'}), n({value: 1, pitch: 'f3'}), n({value: 1, pitch: 'g3'}),
-            n({value: 1, pitch: 'e3'}), n({value: 1, pitch: 'f3'}), n({value: 1, pitch: 'g3'})
+            n({value: 2, pitch: 'e3', slur: 1}), n({value: 4, pitch: 'e3', slur: 1}), r({value: 4}),
+            n({value: 4, pitch: 'c3'}), n({value: 4, pitch: 'c3'}), n({value: 4, pitch: 'c3'}), r()
         ]);
 
         var fourfour = scored.timeSig({value: "4/4", measure: 0});
@@ -221,7 +217,7 @@ const examples = {
         }, [fourfour, system1, system2, system3, system4, trebleLine, bassLine, page0, page1]);
 
         // render it all as a score.
-        return scored.render(score, {voices: [soprano, bass], pages: [1]});
+        return scored.render(score, {voices: [soprano, bass], pages: [0]});
     },
 
     testLegato: function (scored) {