+++ packages/boethius/CHANGELOG.md
@@ -4,3 +4,5 @@ CHANGELOG
 2018-04-24
 ----------
 1. Slur renamed to Legato.
+2. tie utils for drawing slurs and legatos.
+3. Slur draws ties between every note in its group.